Update the onboarding checklist to account for new training schedule
We have recently discussed as a team that we will change our training schedule in order to remove the 10-day requirement for the contingency planning training.  In order to do this, we cannot grant full admin AWS access right away, so we are changing the onboarding checklist to account for this process.

This changeset also adjusts the order of some of the onboarding items and adds a bit more context for the different sections.