Update issue templates for onboarding and offboarding
This changeset updates our onboarding and offboarding checklists and issue templates to make it easier to track this work and consolidate all of the information into one location.  We are shifting these processes to be more open and public while strictly adhering to several compliance controls that we must satisfy.